WIP: set rootdir for NFS
Seems there is a bug in exportfs (specifically the getexportent function).
exportfs tries to resolve symlinks for the exported path to set the export
to the real path. This function does not take into account the rootdir
setting, so the export path /bin tries to resolve to /usr/bin. This then
fails because /srv/nfs/usr/bin does not exist.

Change-Id: I2aa7492c758f343dfcd5305524fe36ba55245a58